chore: standardize view component names
Standardizes view component names to be prefixed with the module name and to be located directly in their respective `views/` directories.

Removes unused MeshDetailIndexView.vue files.
