+++ src/app/diagnostics/views/DiagnosticsDetailView.vue
