+++ src/app/meshes/views/MeshDetailTabsView.vue
