+++ src/app/zone-egresses/views/ZoneEgressDetailView.vue
