+++ src/app/zone-egresses/views/ZoneEgressStatsView.vue
