+++ src/app/zone-ingresses/routes.ts
@@ -6,38 +6,38 @@ export const routes = (prefix = 'ingresses') => {
       {
         path: `${prefix}/:zoneIngress`,
         name: 'zone-ingress-detail-tabs-view',
-        component: () => import('@/app/zone-ingresses/views/item/IndexView.vue'),
+        component: () => import('@/app/zone-ingresses/views/ZoneIngressDetailTabsView.vue'),
         redirect: { name: 'zone-ingress-detail-view' },
         children: [
           {
             path: 'overview',
             name: 'zone-ingress-detail-view',
-            component: () => import('@/app/zone-ingresses/views/item/DetailView.vue'),
+            component: () => import('@/app/zone-ingresses/views/ZoneIngressDetailView.vue'),
           },
           {
             path: 'services',
             name: 'zone-ingress-services-view',
-            component: () => import('@/app/zone-ingresses/views/item/ServicesView.vue'),
+            component: () => import('@/app/zone-ingresses/views/ZoneIngressServicesView.vue'),
           },
           {
             path: 'xds-config',
             name: 'zone-ingress-xds-config-view',
-            component: () => import('@/app/zone-ingresses/views/item/XdsConfigView.vue'),
+            component: () => import('@/app/zone-ingresses/views/ZoneIngressXdsConfigView.vue'),
           },
           {
             path: 'stats',
             name: 'zone-ingress-stats-view',
-            component: () => import('@/app/zone-ingresses/views/item/StatsView.vue'),
+            component: () => import('@/app/zone-ingresses/views/ZoneIngressStatsView.vue'),
           },
           {
             path: 'clusters',
             name: 'zone-ingress-clusters-view',
-            component: () => import('@/app/zone-ingresses/views/item/ClustersView.vue'),
+            component: () => import('@/app/zone-ingresses/views/ZoneIngressClustersView.vue'),
           },
           {
             path: 'config',
             name: 'zone-ingress-config-view',
-            component: () => import('@/app/zone-ingresses/views/item/ConfigView.vue'),
+            component: () => import('@/app/zone-ingresses/views/ZoneIngressConfigView.vue'),
           },
         ],
       },
@@ -53,7 +53,7 @@ export const routes = (prefix = 'ingresses') => {
           meta: {
             module: 'zone-ingresses',
           },
-          component: () => import('@/app/zone-ingresses/views/IndexView.vue'),
+          component: () => import('@/app/zone-ingresses/views/ZoneIngressListView.vue'),
           children: [
             {
               path: ':zoneIngress',
