+++ src/app/zone-ingresses/views/ZoneIngressDetailView.vue
