+++ src/app/zones/views/ZoneConfigView.vue
